build api group && fix relation entity

+++ be/src/main/java/com/teamup/main/service/UserService.java
@@ -1,6 +1,7 @@
 package com.teamup.main.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,52 +11,58 @@ import com.teamup.main.dto.request.UserUpdateRequest;
 import com.teamup.main.exception.AppException;
 import com.teamup.main.exception.ErrorCode;
 import com.teamup.main.mapper.UserMapper;
-import com.teamup.main.model.User;
+import com.teamup.main.model.Users;
 import com.teamup.main.repository.UserRepository;
 
+import lombok.experimental.FieldDefaults;
+
 @Service
+@FieldDefaults(level = lombok.AccessLevel.PRIVATE)
 public class UserService {
     @Autowired
-    private UserRepository userRepository;
+    UserRepository userRepository;
 
     @Autowired
-    private UserMapper userMapper;
+    UserMapper userMapper;
 
-    //
-    // User only
-    //
-    public List<User> getUsersByStudentId(String studentId) {
+    /*
+     * User only
+     */
+    public List<Users> getUsersByStudentId(String studentId) {
         return userRepository.findByStudentIdContainingIgnoreCase(studentId);
     }
 
-    public User updateUser(String userId, UserUpdateRequest request) {
-        User user = getUserById(userId);
+    public Users updateUser(String userId, UserUpdateRequest request) {
+        Users user = getUserById(userId);
         userMapper.updateUser(user, request);
         return userRepository.save(user);
     }
-    
+
     public void deleteUser(String userId) {
-        userRepository.deleteById(userId);
+        if (getUserById(userId) != null) {
+            userRepository.deleteById(userId);
+        }
+        throw new AppException(ErrorCode.USER_NOT_FOUND);
     }
 
-    //
-    // Admin only
-    //
-    public User createUser(GoogleAccount request) {
-        java.util.Optional<User> userOpt = userRepository.findByEmail(request.getEmail());
+    /**
+     * Admin only
+     */
+    public Users createUser(GoogleAccount request) {
+        Optional<Users> userOpt = userRepository.findByEmail(request.getEmail());
         if (userOpt.isPresent()) {
-            return getUserById(userOpt.get().getUserId());
+            return userOpt.get();
         }
 
-        User user = userMapper.toUser(request);
+        Users user = userMapper.toUser(request);
         return userRepository.save(user);
     }
 
-    public List<User> getUsers() {
+    public List<Users> getUsers() {
         return userRepository.findAll();
     }
 
-    public User getUserById(String userId) {
+    public Users getUserById(String userId) {
         return userRepository.findById(userId).orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
     }
 }
